9-12-2019

+++ atividade002/atividade_pratica_2/app/Http/Controllers/SubjectController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Requet;
 use App\Subject;
 use Illuminate\Http\Request;
 use Mockery\Matcher\Subset;
+use Illuminate\Support\Facades\Auth;
 
 class SubjectController extends Controller
 {
@@ -13,6 +15,10 @@ class SubjectController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     public function index()
     {
         //
@@ -25,7 +31,11 @@ class SubjectController extends Controller
      */
     public function create()
     {
-        return view('subjects.create');
+        if (Auth::user()->type == 1) {
+            return view('subjects.create');
+        } else {
+            return redirect()->route('protocols');
+        }
     }
 
     /**
@@ -36,12 +46,16 @@ class SubjectController extends Controller
      */
     public function store(Request $request)
     {
-        $subject= new Subject;
-        $subject->fill($request->all());
-        $subject->save();
-        // Subject::create($request->all());
-
-        return redirect()->route('subjects.show', $subject);
+        if (Auth::user()->type == 1) {
+            $subject = new Subject;
+            $subject->fill($request->all());
+            $subject->save();
+            // Subject::create($request->all());
+            session()->flash('mensagem', 'subject successfully created!');
+            return redirect()->route('subjects.show', $subject);
+        } else {
+            return redirect()->route('protocols');
+        }
     }
 
     /**
@@ -63,7 +77,11 @@ class SubjectController extends Controller
      */
     public function edit(Subject $subject)
     {
-        return view('subjects.edit', ['subject' => $subject]);
+        if (Auth::user()->type == 1) {
+            return view('subjects.edit', ['subject' => $subject]);
+        } else {
+            return redirect()->route('protocols');
+        }
     }
 
     /**
@@ -75,12 +93,13 @@ class SubjectController extends Controller
      */
     public function update(Request $request, Subject $subject)
     {
-        $subject->fill($request->all());
-        //Persiste no BD  
+        if (Auth::user()->type == 1) {
+            $subject->fill($request->all());
+            //Persiste no BD  
 
-        $subject->save();
-        session()->flash('mensagem', 'subject atualizada com sucesso!');
-
+            $subject->save();
+            session()->flash('mensagem', 'subject updated successfully!');
+        }
         return redirect()->route('subjects.show', $subject);
     }
 
@@ -92,8 +111,17 @@ class SubjectController extends Controller
      */
     public function destroy(Subject $subject)
     {
-        $subject->delete();
-        session()->flash('mensagem', 'subject excluida com sucesso!');
-        return redirect()->route('home');
+        if (Auth::user()->type == 1) {
+            $p = Requet::where('subject_id', '=', $subject->id)->get();
+            // dd($p);
+            if ($p->isEmpty()) {
+                $subject->delete();
+                session()->flash('mensagem', 'subject deleted successfully!');
+            } else {
+                session()->flash('mensagem', 'subject linked to request and cannot be deleted!');
+                return redirect()->route('protocols');
+            }
+        }
+        return redirect()->route('protocols');
     }
 }
